+++ include/linux/sysv_fs_sb.h
@@ -50,21 +50,21 @@ struct sysv_sb_info {
 	   different superblock layout. */
 	char *         s_sbd1;		/* entire superblock data, for part 1 */
 	char *         s_sbd2;		/* entire superblock data, for part 2 */
-	unsigned short *s_sb_fic_count;	/* pointer to s_sbd->s_ninode */
-        unsigned short *s_sb_fic_inodes; /* pointer to s_sbd->s_inode */
-	unsigned short *s_sb_total_free_inodes; /* pointer to s_sbd->s_tinode */
-	unsigned short *s_sb_flc_count;	/* pointer to s_sbd->s_nfree */
-	unsigned long  *s_sb_flc_blocks; /* pointer to s_sbd->s_free */
-	unsigned long  *s_sb_total_free_blocks;/* pointer to s_sbd->s_tfree */
-	unsigned long  *s_sb_time;	/* pointer to s_sbd->s_time */
-	unsigned long  *s_sb_state;	/* pointer to s_sbd->s_state, only FSTYPE_SYSV */
+	u16            *s_sb_fic_count;	/* pointer to s_sbd->s_ninode */
+        u16            *s_sb_fic_inodes; /* pointer to s_sbd->s_inode */
+	u16            *s_sb_total_free_inodes; /* pointer to s_sbd->s_tinode */
+	u16            *s_sb_flc_count;	/* pointer to s_sbd->s_nfree */
+	u32	       *s_sb_flc_blocks; /* pointer to s_sbd->s_free */
+	u32            *s_sb_total_free_blocks;/* pointer to s_sbd->s_tfree */
+	u32            *s_sb_time;	/* pointer to s_sbd->s_time */
+	u32            *s_sb_state;	/* pointer to s_sbd->s_state, only FSTYPE_SYSV */
 	/* We keep those superblock entities that don't change here;
 	   this saves us an indirection and perhaps a conversion. */
-	unsigned long  s_firstinodezone; /* index of first inode zone */
-	unsigned long  s_firstdatazone;	/* same as s_sbd->s_isize */
-	unsigned long  s_ninodes;	/* total number of inodes */
-	unsigned long  s_ndatazones;	/* total number of data zones */
-	unsigned long  s_nzones;	/* same as s_sbd->s_fsize */
+	u32            s_firstinodezone; /* index of first inode zone */
+	u32            s_firstdatazone;	/* same as s_sbd->s_isize */
+	u32            s_ninodes;	/* total number of inodes */
+	u32            s_ndatazones;	/* total number of data zones */
+	u32            s_nzones;	/* same as s_sbd->s_fsize */
 };
 /* The fields s_ind_per_block_2_1, s_toobig_block are currently unused. */
